refactor(otel): consolidate bridge to 4 files, drop historicalReplayed
Merge config + mux + stream + util into engine.go and main.go:
  main.go      (lifecycle, signal handling, timing helpers, counters)
  engine.go    (config + discovery + Connect clients + mux + stream)
  writer.go    (ClickHouse batched writer — unchanged)
  checkpoint.go (state persistence — unchanged)

Also: always pass ReturnAllLogs=true on every stream connect. The bridge
checkpoint already filters out replayed batches, so disconnect-window
losses are gone for free. Drops the historicalReplayed flag (codex
non-blocking note from final review).
